+++ lib/glimmer/libui/control_proxy/area_proxy.rb
@@ -36,17 +36,17 @@ module Glimmer
           attr_accessor :current_area_draw_params
         end
         
-        LISTENERS = ['on_draw', 'on_mouse_event', 'on_mouse_down', 'on_mouse_up', 'on_mouse_drag_start', 'on_mouse_drag', 'on_mouse_drop', 'on_mouse_crossed', 'on_mouse_enter', 'on_mouse_exit']
+        LISTENERS = ['on_draw', 'on_mouse_event', 'on_mouse_move', 'on_mouse_down', 'on_mouse_up', 'on_mouse_drag_start', 'on_mouse_drag', 'on_mouse_drop', 'on_mouse_crossed', 'on_mouse_enter', 'on_mouse_exit', 'on_drag_broken']
         LISTENER_ALIASES = {
           on_drawn: 'on_draw',
-          on_mouse_move: 'on_mouse_event',
-          on_mouse_moved: 'on_mouse_event',
+          on_mouse_moved: 'on_mouse_move',
           on_mouse_drag_started: 'on_mouse_drag_start',
           on_mouse_dragged: 'on_mouse_drag',
           on_mouse_dropped: 'on_mouse_drop',
           on_mouse_cross: 'on_mouse_crossed',
           on_mouse_entered: 'on_mouse_enter',
           on_mouse_exited: 'on_mouse_exit',
+          on_drag_break: 'on_drag_broken',
         }
         
         include Glimmer::FiddleConsumer
@@ -88,6 +88,7 @@ module Glimmer
             area_mouse_event = ::LibUI::FFI::AreaMouseEvent.new(area_mouse_event)
             area_mouse_event = area_mouse_event_hash(area_mouse_event)
             on_mouse_event.each { |listener| listener.call(area_mouse_event)}
+            on_mouse_move.each { |listener| listener.call(area_mouse_event)} if area_mouse_event[:x].between?(0, area_mouse_event[:area_width]) && area_mouse_event[:y].between?(0, area_mouse_event[:area_height])
             unless @last_area_mouse_event.nil?
               on_mouse_down.each { |listener| listener.call(area_mouse_event)} if area_mouse_event[:down] > 0 && @last_area_mouse_event[:down] == 0
               on_mouse_up.each { |listener| listener.call(area_mouse_event)} if area_mouse_event[:up] > 0 && @last_area_mouse_event[:up] == 0
@@ -106,7 +107,9 @@ module Glimmer
               on_mouse_enter.each { |listener| listener.call(left) }
             end
           end
-          @area_handler.DragBroken   = fiddle_closure_block_caller(0, [0]) {}
+          @area_handler.DragBroken   = fiddle_closure_block_caller(0, [1, 1]) do |_, _|
+            on_drag_broken.each { |listener| listener.call }
+          end
           @area_handler.KeyEvent     = fiddle_closure_block_caller(0, [0]) {}
         end